Begin restructuring to support multiple lobbies
